수정 사항 추가

+++ src/main/java/com/example/demo/service/BrandService.java
@@ -2,7 +2,9 @@ package com.example.demo.service;
 
 import com.example.demo.dto.BrandInfoDto;
 import com.example.demo.dto.BrandReqDto;
+import com.example.demo.exception.CustomException;
 import com.example.demo.mapper.BrandMapper;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 
 import java.util.ArrayList;
@@ -21,7 +23,11 @@ public class BrandService {
 
     public BrandInfoDto getBrand(Integer brandCode)
     {
-        return brandMapper.getBrand(brandCode);
+        BrandInfoDto brandInfo = brandMapper.getBrand(brandCode);
+        if (brandInfo == null)
+            throw new CustomException("존재하지 않는 브랜드 코드입니다. : " + brandCode, HttpStatus.NOT_FOUND);
+
+        return brandInfo;
     }
 
     public List<Integer> addBrand(List<BrandReqDto> req)
@@ -40,11 +46,19 @@ public class BrandService {
 
     public void setBrand(BrandReqDto req)
     {
+        BrandInfoDto brandInfo = brandMapper.getBrand(req.getBrandCode());
+        if (brandInfo == null)
+            throw new CustomException("존재하지 않는 브랜드 코드입니다. : " + req.getBrandCode(), HttpStatus.NOT_FOUND);
+
         brandMapper.setBrand(req);
     }
 
     public void deleteBrand(Integer brandCode)
     {
+        BrandInfoDto brandInfo = brandMapper.getBrand(brandCode);
+        if (brandInfo == null)
+            throw new CustomException("존재하지 않는 브랜드 코드입니다. : " + brandCode, HttpStatus.NOT_FOUND);
+
         brandMapper.deleteBrand(brandCode);
     }
 }
