+++ src/main/java/com/example/demo/service/ItemService.java
@@ -3,7 +3,9 @@ package com.example.demo.service;
 import com.example.demo.dto.ItemInfoDto;
 import com.example.demo.dto.ItemReqDto;
 import com.example.demo.dto.ItemResDto;
+import com.example.demo.exception.CustomException;
 import com.example.demo.mapper.ItemMapper;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 
 import java.math.BigInteger;
@@ -23,7 +25,12 @@ public class ItemService {
 
     public ItemInfoDto getItem(BigInteger itemNo)
     {
-        return itemMapper.getItem(itemNo);
+        ItemInfoDto item = itemMapper.getItem(itemNo);
+
+        if (item == null)
+            throw new CustomException("존재하지 않는 상품입니다 : " + itemNo, HttpStatus.NOT_FOUND);
+
+        return item;
     }
 
     public List<ItemResDto> addItem(List<ItemReqDto> req)
@@ -45,11 +52,24 @@ public class ItemService {
 
     public void setItem(ItemReqDto req)
     {
+        validationService.validateCategoryCode(req.getCategoryCode());
+        validationService.validateBrandCode(req.getBrandCode());
+
+        ItemInfoDto item = itemMapper.getItem(req.getItemNo());
+
+        if (item == null)
+            throw new CustomException("존재하지 않는 상품입니다 : " + req.getItemNo(), HttpStatus.NOT_FOUND);
+
         itemMapper.setItem(req);
     }
 
     public void deleteItem(BigInteger itemNo)
     {
+        ItemInfoDto item = itemMapper.getItem(itemNo);
+
+        if (itemNo == null)
+            throw new CustomException("존재하지 않는 상품입니다 : " + itemNo, HttpStatus.NOT_FOUND);
+
         itemMapper.deleteItem(itemNo);
     }
